chore: time field BigInt field로 변경

+++ ai_model/models.py
@@ -10,31 +10,32 @@ class YouTubeURL(models.Model):
 class YouTubeInfo(models.Model):
     url_id = models.ForeignKey("YouTubeURL", on_delete=models.CASCADE, db_column='url_id')
     title = models.CharField(help_text="YouTube Title", max_length=50)
-    length = models.IntegerField(help_text="YouTube Length")
+    length = models.BigIntegerField(help_text="YouTube Length")
     def __str__(self):
         return str(self.url_id) + " : " + self.title
 
 class YouTubeCaption(models.Model):
     url_id = models.ForeignKey("YouTubeURL", on_delete=models.CASCADE, db_column='url_id')
-    start_time = models.DecimalField(max_digits=20, decimal_places=4, blank=True, null=True)
-    end_time = models.DecimalField(max_digits=20, decimal_places=4, blank=True, null=True)
+    start_time = models.BigIntegerField(blank=True, null=True)
+    end_time = models.BigIntegerField(blank=True, null=True)
     text = models.TextField(blank=True, null=True)
     def __str__(self):
         return f"{str(self.url_id)} : [{self.start_time:6.2f} ~ {self.end_time:6.2f}] {self.text}"
 
 class OCRResult(models.Model):
     url_id = models.ForeignKey("YouTubeURL", on_delete=models.CASCADE, db_column='url_id')
-    time = models.DecimalField(max_digits=20, decimal_places=4, blank=True, null=True)
+    time = models.BigIntegerField(blank=True, null=True)
     text = models.TextField(blank=True, null=True)
     conf = models.DecimalField(max_digits=7, decimal_places=6, blank=True, null=True)
-    # bbox 는 나중에 추가
+    # bbox = models.TextField(blank=True, null=True)
+    # size = models.TextField(blank=True, null=True)
     def __str__(self):
         return f"{str(self.url_id)} : [{self.time:.2f}] {self.text} ({self.conf:.4f})"
 
 class STTResult(models.Model):
     url_id = models.ForeignKey("YouTubeURL", on_delete=models.CASCADE, db_column='url_id')
-    start_time = models.DecimalField(max_digits=20, decimal_places=4, blank=True, null=True)
-    end_time = models.DecimalField(max_digits=20, decimal_places=4, blank=True, null=True)
+    start_time = models.BigIntegerField(blank=True, null=True)
+    end_time = models.BigIntegerField(blank=True, null=True)
     text = models.TextField(blank=True, null=True)
     def __str__(self):
         return f"{str(self.url_id)} : [{self.start_time:6.2f} ~ {self.end_time:6.2f}] {self.text}"
